+++ new-player/src/main/java/net/newpipe/newplayer/data/NewPlayerException.kt
@@ -18,7 +18,7 @@
  * along with NewPlayer.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.newpipe.newplayer.utils
+package net.newpipe.newplayer.data
 
 class NewPlayerException : Exception {
     constructor(message: String) : super(message)
