+++ new-player/src/main/java/net/newpipe/newplayer/data/VideoSize.kt
@@ -19,7 +19,7 @@
  */
 
 
-package net.newpipe.newplayer.utils
+package net.newpipe.newplayer.data
 
 internal data class VideoSize(
     val width: Int,
