+++ new-player/src/main/java/net/newpipe/newplayer/ui/common/PlaylistControllButtons.kt
@@ -35,8 +35,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.media3.common.util.UnstableApi
 import net.newpipe.newplayer.R
 import net.newpipe.newplayer.RepeatMode
-import net.newpipe.newplayer.model.NewPlayerUIState
-import net.newpipe.newplayer.model.InternalNewPlayerViewModel
+import net.newpipe.newplayer.uiModel.NewPlayerUIState
+import net.newpipe.newplayer.uiModel.InternalNewPlayerViewModel
 
 @OptIn(UnstableApi::class)
 @Composable
