+++ new-player/src/main/java/net/newpipe/newplayer/ui/selection_ui/ChapterItem.kt
@@ -47,8 +47,8 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import net.newpipe.newplayer.utils.Chapter
-import net.newpipe.newplayer.utils.NewPlayerException
+import net.newpipe.newplayer.data.Chapter
+import net.newpipe.newplayer.data.NewPlayerException
 import net.newpipe.newplayer.R
 import net.newpipe.newplayer.ui.theme.VideoPlayerTheme
 import net.newpipe.newplayer.ui.common.Thumbnail
