+++ new-player/src/main/java/net/newpipe/newplayer/ui/selection_ui/StreamItem.kt
@@ -62,7 +62,7 @@ import androidx.compose.ui.unit.sp
 import androidx.media3.common.MediaItem
 import androidx.media3.common.util.UnstableApi
 import net.newpipe.newplayer.R
-import net.newpipe.newplayer.model.NewPlayerUIState
+import net.newpipe.newplayer.uiModel.NewPlayerUIState
 import net.newpipe.newplayer.ui.videoplayer.CONTROLLER_UI_BACKGROUND_COLOR
 import net.newpipe.newplayer.ui.theme.VideoPlayerTheme
 import net.newpipe.newplayer.ui.common.ReorderHapticFeedback
