+++ test-app/src/main/java/net/newpipe/newplayer/testapp/MainActivity.kt
@@ -36,9 +36,9 @@ import kotlinx.coroutines.launch
 import net.newpipe.newplayer.ActivityBrainSlug
 import net.newpipe.newplayer.NewPlayer
 import net.newpipe.newplayer.PlayMode
-import net.newpipe.newplayer.model.NewPlayerViewModel
-import net.newpipe.newplayer.model.NewPlayerViewModelImpl
-import net.newpipe.newplayer.model.UIModeState
+import net.newpipe.newplayer.uiModel.NewPlayerViewModel
+import net.newpipe.newplayer.uiModel.NewPlayerViewModelImpl
+import net.newpipe.newplayer.uiModel.UIModeState
 import net.newpipe.newplayer.testapp.databinding.ActivityMainBinding
 import net.newpipe.newplayer.ui.ContentScale
 import javax.inject.Inject
