+++ test-app/src/main/java/net/newpipe/newplayer/testapp/TestMediaRepository.kt
@@ -12,14 +12,14 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import net.newpipe.newplayer.utils.Chapter
+import net.newpipe.newplayer.data.Chapter
 import net.newpipe.newplayer.MediaRepository
 import net.newpipe.newplayer.RepoMetaInfo
-import net.newpipe.newplayer.utils.AudioStreamTrack
-import net.newpipe.newplayer.utils.LanguageIdentifier
-import net.newpipe.newplayer.utils.Stream
-import net.newpipe.newplayer.utils.Subtitle
-import net.newpipe.newplayer.utils.VideoStreamTrack
+import net.newpipe.newplayer.data.AudioStreamTrack
+import net.newpipe.newplayer.data.LanguageIdentifier
+import net.newpipe.newplayer.data.Stream
+import net.newpipe.newplayer.data.Subtitle
+import net.newpipe.newplayer.data.VideoStreamTrack
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.Response
